refactor(mandate): mechanically split types into bounded modules (wave18 step2)
