+++ crates/assay-evidence/src/mandate/types/schema.rs
@@ -1,0 +1,8 @@
+/// Mandate payload type for type confusion prevention.
+pub const MANDATE_PAYLOAD_TYPE: &str = "application/vnd.assay.mandate+json;v=1";
+
+/// Mandate used event payload type.
+pub const MANDATE_USED_PAYLOAD_TYPE: &str = "application/vnd.assay.mandate.used+json;v=1";
+
+/// Mandate revoked event payload type.
+pub const MANDATE_REVOKED_PAYLOAD_TYPE: &str = "application/vnd.assay.mandate.revoked+json;v=1";
