+++ crates/assay-evidence/src/mandate/types/serde.rs
@@ -1,0 +1,3 @@
+pub(crate) fn is_false(b: &bool) -> bool {
+    !*b
+}
